+++ spec/dummy_app/config/application.rb
@@ -21,13 +21,6 @@ module Dummy
     config.active_support.test_order = :sorted
     config.secret_key_base = "A fox regularly kicked the screaming pile of biscuits."
 
-    # In rails >= 6.0, "`.represent_boolean_as_integer=` is now always true,
-    # so setting this is deprecated and will be removed in Rails 6.1."
-    if ::ENV["DB"] == "sqlite" &&
-        ::Gem::Requirement.new("~> 5.2").satisfied_by?(::Rails.gem_version)
-      config.active_record.sqlite3.represent_boolean_as_integer = true
-    end
-
     # `use_yaml_unsafe_load` was added in 7.0.3.1, will be removed in 7.1.0?
     if ::ActiveRecord.respond_to?(:use_yaml_unsafe_load)
       ::ActiveRecord.use_yaml_unsafe_load = false